build path reconfig to fix failing test

+++ src/domain/TransactionTest.java
@@ -1,8 +1,8 @@
 package domain;
 
 import static org.junit.Assert.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.Assert.assertTrue;
-import static org.junit.jupiter.api.Assertions.*;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -24,7 +24,7 @@ class TransactionTest {
 	
 	@Test
 	public void testGetAmount() {
-		int result = 100;
+		double result = 100.00;
 		assertEquals(result, classUnderTest.getAmount());
 	}
